Fixed non-working browser action.
The whitelist is now properly saved. I also simplified the whitelist
so that it merely stores domains instead of regex. The current page
has also been changed to just save the domain. However, it currently
has issues with the fact that some site have nested pages that cause
the domain to be incorrectly recognized.