Extension no longer blanks user agent.
This was causing issues with Google searches (which assume a blank
UA means mobile browser?). Also formatted background script to match
the other scripts. Finally, the unnecessary background property was
removed -- that just allows the script to run in the background more
than necessary (ie, when the browser isn't even running).