Fixed error messages for uninitialized whitelist.
If the whitelist hasn't yet been set (ie, first run and haven't yet
added anything to the whitelist), you get error messages about
attempting to access non-existant parameters, since the whitelist
is undefined. To prevent that, the first time the extension is loaded
(probably on browser startup), we check if the whitelist is empty and
set it as an empty array if it is.