I'm wrong, JS does allow getters and setters.
There's functions for getters and setters that are run when
accessing parameters automatically. That's pretty cool -- didn't
know that. Also a note that some sites must be whitelisted.

In particular, I can't seem to be able to load my gmail inbox
without adding gmail to the whitelist. I can't figure out why (I
started a SO question about it).